Fixed all warnings other than variable length arrays (at least for the console)

+++ console.cpp
@@ -132,8 +132,8 @@ void print_dims(TemplatePointsMessage const& message, std::ostream& ostream)
     auto data = message.homology_dimensions.data();
     ostream << "Dimensions > 0:" << std::endl;
 
-    for (int row = 0; row < shape[0]; row++) {
-        for (int col = 0; col < shape[1]; col++) {
+    for (unsigned long row = 0; row < shape[0]; row++) {
+        for (unsigned long col = 0; col < shape[1]; col++) {
             unsigned dim = data[row * shape[0] + col];
             if (dim > 0) {
                 ostream << "(" << col << ", " << row << ", " << dim << ")" << std::endl;
